Upgrade near-sdk and borsh dependencies (#47)
Update near-sdk to 5.24 and borsh to 1.6 across main contract and v2
fixture. Migrate from deprecated collections module to store module
(IterableMap, LookupSet). Update type annotations for compatibility
with latest SDK (u32 for counts and pagination parameters).

All 51 integration tests pass.

+++ contracts/verified-accounts/tests/integration/initialization_tests.rs
@@ -25,7 +25,7 @@ async fn test_contract_initialization() -> anyhow::Result<()> {
     });
 
     // Verify initial count is 0
-    let count: u64 = contract.view("get_verified_count").await?.json()?;
+    let count: u32 = contract.view("get_verified_count").await?.json()?;
 
     step("Verify initial count is 0", || {
         assert_eq!(count, 0);
